+++ src/components/chart/pieChart/styles.ts
@@ -1,0 +1,4 @@
+export const COLORS = ["#343A6E", "#030035", "#DDF3FE", "#FF9800"];
+
+export const legendNameStyle = { marginLeft: "18px", color: "#030035" };
+export const legendValStyle = { color: "#9FA2B4", marginRight: "10px" };
