+++ src/components/chart/pieChart/types.ts
@@ -1,0 +1,1 @@
+export type PieChartData = { name: string; value: number };
