+++ src/components/chart/pieChart/utils.ts
@@ -1,0 +1,3 @@
+export const percentage = (total: number, value: number) => {
+  return Math.round((value / total) * 100);
+};
